:sparkles: CountdownContext
Created countdown context.
