+++ #4 NLW/ReactJS/moveit-next/src/pages/index.tsx
@@ -5,6 +5,7 @@ import { CompletedChallenges } from "../components/CompletedChallenges";
 import { Countdown } from "../components/Countdown";
 import { ExperienceBar } from "../components/ExperienceBar";
 import { Profile } from "../components/Perfil";
+import { CountdownProvider } from "../contexts/CountdownContext";
 
 import styles from "../styles/pages/Home.module.css";
 
@@ -17,16 +18,18 @@ export default function Home() {
 
             <ExperienceBar />
 
-            <section>
-                <div>
-                    <Profile />
-                    <CompletedChallenges />
-                    <Countdown />
-                </div>
-                <div>
-                    <ChallengeBox />
-                </div>
-            </section>
+            <CountdownProvider>
+                <section>
+                    <div>
+                        <Profile />
+                        <CompletedChallenges />
+                        <Countdown />
+                    </div>
+                    <div>
+                        <ChallengeBox />
+                    </div>
+                </section>
+            </CountdownProvider>
         </div>
     );
 }
